changed variables to fit vlc

+++ ScreenTeamFX/src/modules/Event.java
@@ -12,7 +12,7 @@ public class Event implements Comparable<Event>{
 	 * TODO: should 'time' be from the start of the performance (00:00) or from where the it starts when
 	 * 	the user presses play?
 	 */
-	private int time;
+	private long time;
 	private int timelineid;
 	private Action action;
 	private TimelineMediaObject timelineMediaObject;
@@ -24,7 +24,7 @@ public class Event implements Comparable<Event>{
 	 * @param action
 	 * @param mediaobject	The mediaobject that this event is associated with.
 	 */
-	public Event(int time, int timelineid, Action action, TimelineMediaObject timelineMediaObject) {
+	public Event(long time, int timelineid, Action action, TimelineMediaObject timelineMediaObject) {
 		super();
 		this.time = time;
 		this.timelineid = timelineid;
@@ -32,7 +32,7 @@ public class Event implements Comparable<Event>{
 		this.setTimelineMediaObject(timelineMediaObject);
 	}
 
-	public int getTime() {
+	public long getTime() {
 		return time;
 	}
 
@@ -63,7 +63,7 @@ public class Event implements Comparable<Event>{
 	@Override
 	public int compareTo(Event otherEvent) {
 		
-		int otherTime = otherEvent.getTime();
+		long otherTime = otherEvent.getTime();
 		
 		/**
 		 * If one MediaObject stops and another one starts at the same time and on the same Timeline,
@@ -86,7 +86,7 @@ public class Event implements Comparable<Event>{
 			}
 		}
 		
-		return this.time - otherTime;
+		return (int) (this.time - otherTime);
 		
 		// Descending order
 		// return compareTime - this.time;
